Refactor layout / widget interface
This allows to cut loose quite some dependencies on headers.
