+++ radio/src/gui/colorlcd/widgets/outputs.cpp
@@ -19,6 +19,7 @@
  */
 
 #include "opentx.h"
+#include "widgets_container_implem.h"
 
 #define RECT_BORDER                    1
 #define ROW_HEIGHT                     17
