+++ radio/src/gui/colorlcd/widgets/text.cpp
@@ -19,6 +19,7 @@
  */
 
 #include "opentx.h"
+#include "widgets_container_implem.h"
 
 #define TEXT_WIDGET_DEFAULT_LABEL  'M', 'y', ' ', 'L', 'a', 'b', 'e', 'l' // "My Label"
