+++ radio/src/gui/colorlcd/widgets_container.h
@@ -22,146 +22,63 @@
 #define _WIDGETS_CONTAINER_H_
 
 #include <stdlib.h>
-#include <window.h>
-#include "widget.h"
+#include "libopenui_types.h"
+#include "form.h"
+#include "zone.h"
 
-class WidgetsContainerInterface
+#define WIDGET_NAME_LEN     10
+#define MAX_WIDGET_OPTIONS   5 // Name?
+
+#define MAX_TOPBAR_ZONES     4
+#define MAX_TOPBAR_OPTIONS   1 // just because of VC++ which doesn't like 0-size arrays :(
+
+// Common 'ZoneOptionValue's among all layouts
+enum {
+  LAYOUT_OPTION_TOPBAR = 0,
+  LAYOUT_OPTION_FM,
+  LAYOUT_OPTION_SLIDERS,
+  LAYOUT_OPTION_TRIMS,
+  LAYOUT_OPTION_MIRRORED,
+
+  LAYOUT_OPTION_LAST_DEFAULT=LAYOUT_OPTION_MIRRORED
+};
+
+class Widget;
+class WidgetFactory;
+class LayoutFactory;
+
+struct WidgetPersistentData {
+  ZoneOptionValueTyped options[MAX_WIDGET_OPTIONS] USE_IDX;
+};
+
+struct ZonePersistentData {
+  char widgetName[WIDGET_NAME_LEN];
+  WidgetPersistentData widgetData;
+};
+
+template<int N, int O>
+struct WidgetsContainerPersistentData {
+  ZonePersistentData   zones[N];
+  ZoneOptionValueTyped options[O];
+};
+
+typedef WidgetsContainerPersistentData<MAX_TOPBAR_ZONES, MAX_TOPBAR_OPTIONS> TopBarPersistentData;
+
+class WidgetsContainerInterface: public FormGroup
 {
   public:
+    using FormGroup::FormGroup;
+  
     virtual unsigned int getZonesCount() const = 0;
     virtual rect_t getZone(unsigned int index) const = 0;
     virtual Widget * createWidget(unsigned int index, const WidgetFactory * factory) = 0;
     virtual Widget * getWidget(unsigned int index) = 0;
     virtual void removeWidget(unsigned int index) = 0;
+
+    // this should go away...
+    virtual const LayoutFactory * getFactory() const = 0;
+    virtual void decorate() = 0;
 };
 
-#define WIDGET_NAME_LEN   10
-
-template<int N, int O>
-class WidgetsContainer: public FormGroup, public WidgetsContainerInterface
-{
-  public:
-    struct ZonePersistentData {
-      char widgetName[WIDGET_NAME_LEN];
-      Widget::PersistentData widgetData;
-    };
-
-    struct PersistentData {
-      ZonePersistentData   zones[N];
-      ZoneOptionValueTyped options[O];
-    };
-
-    WidgetsContainer(const rect_t & rect, PersistentData * persistentData):
-      FormGroup(nullptr, rect, FORM_FORWARD_FOCUS),
-      persistentData(persistentData)
-    {
-    }
-
-    Widget * createWidget(unsigned int index, const WidgetFactory * factory) override
-    {
-      if (index >= N)
-        return nullptr;
-
-      // remove old one if existing
-      removeWidget(index);
-
-      Widget * widget = nullptr;
-      if (factory) {
-        strncpy(persistentData->zones[index].widgetName, factory->getName(), sizeof(ZonePersistentData::widgetName));
-        widget = factory->create(this, getZone(index), &persistentData->zones[index].widgetData);
-      }
-      widgets[index] = widget;
-
-      if (widget)
-        widget->attach(this);
-
-      return widget;
-    }
-
-    void removeWidget(unsigned int index) override
-    {
-      if (index >= N)
-        return;
-
-      if (widgets[index])
-        widgets[index]->deleteLater();
-
-      widgets[index] = nullptr;
-      memset(persistentData->zones[index].widgetName, 0, sizeof(ZonePersistentData::widgetName));
-      memset(&persistentData->zones[index].widgetData, 0, sizeof(Widget::PersistentData));
-    }
-
-    Widget * getWidget(unsigned int index) override
-    {
-      if (index < N)
-        return widgets[index];
-
-      return nullptr;
-    }
-
-    virtual void create()
-    {
-      memset(persistentData, 0, sizeof(PersistentData));
-    }
-
-    virtual void load()
-    {
-      unsigned int count = getZonesCount();
-      for (unsigned int i = 0; i < count; i++) {
-
-        // remove old widget
-        if (widgets[i]) {
-          widgets[i]->deleteLater();
-          widgets[i] = nullptr;
-        }
-
-        // and load new one if required
-        if (persistentData->zones[i].widgetName[0]) {
-          char name[WIDGET_NAME_LEN + 1];
-          memset(name, 0, sizeof(name));
-          strncpy(name, persistentData->zones[i].widgetName, WIDGET_NAME_LEN);
-          widgets[i] = loadWidget(name, this, getZone(i), &persistentData->zones[i].widgetData);
-        }
-      }
-    }
-
-    inline ZoneOptionValue * getOptionValue(unsigned int index) const
-    {
-      return &persistentData->options[index].value;
-    }
-
-    inline void setOptionValue(unsigned int index, const ZoneOptionValue& value)
-    {
-      persistentData->options[index].value = value;
-    }
-
-    unsigned int getZonesCount() const override = 0;
-
-    rect_t getZone(unsigned int index) const override = 0;
-
-    virtual void background()
-    {
-      for (int i = 0; i < N; i++) {
-        if (widgets[i]) {
-          widgets[i]->background();
-        }
-      }
-    }
-
-    void updateZones()
-    {
-      for (int i = 0; i < N; i++) {
-        if (widgets[i]) {
-          auto zone = getZone(i);
-          widgets[i]->setRect(zone);
-          widgets[i]->setInnerHeight(zone.h);
-        }
-      }
-    }
-  
-  protected:
-    PersistentData * persistentData;
-    Widget * widgets[N] = {};
-};
 
 #endif // _WIDGETS_CONTAINER_H_
